Implemented saving player class to file
Previously the player was asked each time what class they wanted to play as. Now, the class is saved to the first line of data.txt as a class code (1- Warrior, 2- Rogue, etc). This class code is read when the game begins if data.txt exists, and sets the player class accordingly.