Build on the command line on all OS
With cmake 3.13 release the `-S <source_dir>` and `-B <build_dir>`
parameter were added.

https://cmake.org/cmake/help/v3.13/release/3.13.html#command-line

With this change the need to create a build directory and then `cd` into
it is no more. Just `cmake -S . -B build` in the source directory and we
have an out of source build.

This also works with Visual Studio if the right generator expression is
passed to it.

One can still open the solution file in Visual Studio.